+++ backend/src/layer/trigger/graphql/mod.rs
@@ -7,4 +7,4 @@ pub mod server;
 mod types;
 pub mod util;
 
-mod test;
+mod tests;
